+++ bindings/dotnet/src/Pamoja.Routing/Routing.cs
@@ -1,4 +1,3 @@
-using Pamoja.Core;
 using Pamoja.Native.Interop;
 
 namespace Pamoja.Routing;